Move tests to WebAssembly-only scenarios
